refactor: Player
